+++ e2e/specs/AntarcticBbox.spec.ts
@@ -1,7 +1,11 @@
-import AxeBuilder from '@axe-core/playwright';
-import { expect, test } from '@playwright/test';
+import { test } from '@playwright/test';
 
-import { runAccessibilityCheck, waitForMapReady } from '../config/test.utils';
+import {
+  runAccessibilityCheck,
+  testSnapshot,
+  waitForMapReady,
+  waitForSceneReady,
+} from '../config/test.utils';
 
 const bboxes = [
   [-180.0, -90.0, 180.0, -60.0],
@@ -11,27 +15,23 @@ const bboxes = [
   [-180.0, -90.0, 180.0, -50.0],
   [140.0, -90.0, -60.0, -50.0],
   [140.0, -80.0, -60.0, -50.0],
+  [
+    [-180.0, -90.0, 180.0, -80.0],
+    [60.0, -75.0, 180.0, -60.0],
+    [-180.0, -58.0, 180.0, -50.0],
+  ],
 ];
 
 test.describe.parallel('Antarctic Bounding Boxes', () => {
   for (const bbox of bboxes) {
     test.describe(`bbox=${bbox}`, () => {
       test.beforeEach(async ({ page }) => {
-        await page.goto(`/?bbox=[${bbox.join(',')}]`);
+        await page.goto(`/?bbox=${JSON.stringify(bbox)}`);
         await waitForMapReady(page);
       });
 
       test('snapshot', async ({ page }) => {
-        // Wait for the ArcGIS map view to be ready
-        await page.waitForSelector('arcgis-map', { state: 'visible' });
-
-        // Wait for the map to finish updating
-        await page.waitForSelector('arcgis-map:not([updating])', {
-          state: 'visible',
-          timeout: 20000,
-        });
-
-        await expect(page).toHaveScreenshot(`bbox-${bbox.join('-')}.png`, { fullPage: true });
+        await testSnapshot(page, `bbox-${bbox.join('-')}`);
       });
 
       test('should not have any automatically detectable accessibility issues', async ({
@@ -44,22 +44,17 @@ test.describe.parallel('Antarctic Bounding Boxes', () => {
 });
 
 test.describe('Antarctic Bounding Box with globe overview', () => {
-  test('should not have any automatically detectable accessibility issues', async ({ page }) => {
+  test.beforeEach(async ({ page }) => {
     await page.goto('/?bbox=[-180,-90,180,-60]&globe-overview=true');
+    await waitForMapReady(page);
+    await waitForSceneReady(page);
+  });
 
-    await page.waitForSelector('arcgis-map', { state: 'visible' });
+  test('snapshot', async ({ page }) => {
+    await testSnapshot(page, 'bbox-antarctic-globe-overview');
+  });
 
-    await page.waitForSelector('arcgis-map:not([updating])', {
-      state: 'visible',
-      timeout: 20000,
-    });
-
-    // wait 5 seconds
-    await page.waitForTimeout(5000);
-
-    await expect(page).toHaveScreenshot('bbox-antarctic-globe-overview.png', { fullPage: true });
-
-    const accessibilityScanResults = await new AxeBuilder({ page }).analyze();
-    expect(accessibilityScanResults.violations).toEqual([]);
+  test('should not have any automatically detectable accessibility issues', async ({ page }) => {
+    await runAccessibilityCheck(page);
   });
 });