+++ src/components/App/App.tsx
@@ -5,13 +5,6 @@ import { useMapParams } from '@/lib/hooks/useMapParams';
 
 import { Map } from '../Map/Map';
 
-function convertEmptyStringToBooleanPresence(
-  value: string | boolean | undefined,
-): boolean | undefined {
-  if (value === '' || typeof value === 'string') return true;
-  return value;
-}
-
 export function App() {
   const {
     // View parameters
@@ -32,6 +25,9 @@ export function App() {
     // Asset parameters
     assetId,
     assetForcePopup,
+
+    // Overlays
+    showGraticule,
   } = useMapParams();
 
   return (
@@ -40,14 +36,15 @@ export function App() {
         initialAssetId={assetId}
         initialCenter={centre}
         initialZoom={zoom}
-        initialBbox={bbox as [number, number, number, number]}
-        bboxForceRegionalExtent={convertEmptyStringToBooleanPresence(bboxForceRegionalExtent)}
+        initialBbox={bbox}
+        bboxForceRegionalExtent={bboxForceRegionalExtent}
         initialScale={scale}
-        showGlobeOverview={convertEmptyStringToBooleanPresence(showGlobeOverview)}
-        showZoomButton={convertEmptyStringToBooleanPresence(showZoomButton)}
-        showResetButton={convertEmptyStringToBooleanPresence(showResetButton)}
-        showFullscreenButton={convertEmptyStringToBooleanPresence(showFullscreenButton)}
-        initialShowAssetPopup={convertEmptyStringToBooleanPresence(assetForcePopup)}
+        showGraticule={showGraticule}
+        showGlobeOverview={showGlobeOverview}
+        showZoomButton={showZoomButton}
+        showResetButton={showResetButton}
+        showFullscreenButton={showFullscreenButton}
+        initialShowAssetPopup={assetForcePopup}
       />
     </GeometryToolsLoader>
   );