+++ src/components/Map/Map.tsx
@@ -3,6 +3,7 @@ import '@arcgis/map-components/components/arcgis-placement';
 import React from 'react';
 
 import { ArcMapView } from '@/lib/arcgis/components/ArcView/ArcMapView';
+import { GraticuleLayer } from '@/lib/arcgis/customlayers/GraticuleLayer/GraticuleLayer';
 
 import { Globe } from '../Globe';
 import LoadingScrim from '../LoadingScrim';
@@ -11,12 +12,14 @@ import HomeControl from '../map-controls/HomeControl';
 import ScaleControl from '../map-controls/ScaleControl/ScaleControl';
 import ZoomControl from '../map-controls/ZoomControl';
 import { useMapInitialization } from './hooks/useMapInitialization';
+import { BBox } from './utils/bboxUtils';
+
 interface MapProps {
   // View parameters
   initialZoom?: number;
   initialScale?: number;
   initialCenter?: [number, number];
-  initialBbox?: [number, number, number, number];
+  initialBbox?: BBox[];
   bboxForceRegionalExtent?: boolean;
 
   // UI Controls
@@ -30,6 +33,9 @@ interface MapProps {
   // Asset parameters
   initialAssetId?: string;
   initialShowAssetPopup?: boolean;
+
+  // Overlays
+  showGraticule?: boolean;
 }
 
 export function Map({
@@ -43,6 +49,7 @@ export function Map({
   showZoomButton,
   showResetButton,
   showFullscreenButton,
+  showGraticule,
   initialShowAssetPopup,
 }: MapProps) {
   const [viewPoint, setViewPoint] = React.useState<__esri.Viewpoint | undefined>(undefined);
@@ -75,6 +82,7 @@ export function Map({
         scale={initialScale}
         zoom={initialZoom}
       >
+        {showGraticule && <GraticuleLayer />}
         <arcgis-placement position="top-left">
           <div className="flex flex-col gap-4">
             {showZoomButton && <ZoomControl />}