+++ src/lib/hooks/useMapParams.ts
@@ -1,8 +1,10 @@
+import { bboxParamToArray } from '@/components/Map/utils/bboxUtils';
 import {
   DEFAULT_REGIONAL_EXTENT,
   DEFAULT_SHOW_ASSET_POPUP,
   DEFAULT_SHOW_FULLSCREEN_BUTTON,
   DEFAULT_SHOW_GLOBE_OVERVIEW,
+  DEFAULT_SHOW_GRATICULE,
   DEFAULT_SHOW_RESET_BUTTON,
   DEFAULT_SHOW_ZOOM_BUTTON,
 } from '@/lib/config/mapParamDefaults';
@@ -35,6 +37,9 @@ export function useMapParams() {
     // Asset parameters
     'asset-id': assetId,
     'asset-force-popup': assetForcePopup = DEFAULT_SHOW_ASSET_POPUP,
+
+    // Overlays
+    graticule: showGraticule = DEFAULT_SHOW_GRATICULE,
   } = Route.useSearch();
 
   return {
@@ -42,8 +47,8 @@ export function useMapParams() {
     zoom,
     scale,
     centre,
-    bbox,
-    bboxForceRegionalExtent,
+    bbox: bbox ? bboxParamToArray(bbox) : undefined,
+    bboxForceRegionalExtent: convertEmptyStringParamToBooleanPresence(bboxForceRegionalExtent),
 
     // UI Controls
     showZoomButton: convertEmptyStringParamToBooleanPresence(showZoomButton),
@@ -56,5 +61,8 @@ export function useMapParams() {
     // Asset parameters
     assetId,
     assetForcePopup: convertEmptyStringParamToBooleanPresence(assetForcePopup),
+
+    // Overlays
+    showGraticule: convertEmptyStringParamToBooleanPresence(showGraticule),
   };
 }